rename helpers to notebook-helpers
